pyproject.toml
[project]
name = "pixel-ai"
version = "0.1.2"
description = "Offline Raspberry Pi-optimized LLM for Emotion Mirror Project"
authors = [{ name="Your Name" }]
dependencies = [
    "torch",
    "transformers",
    "llama-cpp-python",
    "click",
    "numpy",
    "requests",
    "rich"
]

[project.scripts]
pixel-ai = "pixel_ai.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["pixel_ai*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
